Extract common ResolvableGolangPackage
