+++ src/main/java/com/github/blindpirate/gogradle/core/LocalDirectoryGolangPackage.java
@@ -1,31 +1,25 @@
 package com.github.blindpirate.gogradle.core;
 
-import com.github.blindpirate.gogradle.util.StringUtils;
-
-import java.io.File;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.Optional;
 
-public class LocalDirectoryGolangPackage extends GolangPackage {
+public class LocalDirectoryGolangPackage extends ResolvableGolangPackage {
 
-    private String rootPathString;
+    private String dir;
 
-    private File dir;
-
-    private LocalDirectoryGolangPackage(Path rootPath, Path path, File dir) {
-        super(path);
-        rootPathString = StringUtils.toUnixString(rootPath);
+    private LocalDirectoryGolangPackage(Path rootPath, Path path, String dir) {
+        super(rootPath, path);
         this.dir = dir;
     }
 
-    public String getRootPathString() {
-        return rootPathString;
+    public String getDir() {
+        return dir;
     }
 
     @Override
     protected Optional<GolangPackage> longerPath(Path packagePath) {
-        return Optional.of(of(rootPathString, getPathString(), dir));
+        return Optional.of(of(getRootPathString(), getPathString(), dir));
     }
 
     @Override
@@ -33,11 +27,11 @@ public class LocalDirectoryGolangPackage extends GolangPackage {
         return Optional.of(IncompleteGolangPackage.of(packagePath));
     }
 
-    public static LocalDirectoryGolangPackage of(Path rootPath, Path path, File dir) {
+    public static LocalDirectoryGolangPackage of(Path rootPath, Path path, String dir) {
         return new LocalDirectoryGolangPackage(rootPath, path, dir);
     }
 
-    public static LocalDirectoryGolangPackage of(String rootPath, String path, File dir) {
+    public static LocalDirectoryGolangPackage of(String rootPath, String path, String dir) {
         return new LocalDirectoryGolangPackage(Paths.get(rootPath), Paths.get(path), dir);
     }
 }
